Refactor project: - `light-elastic` now contains the core implementation of ES materialization. Move ES TCP materializer to new `lightql-elastic-tcp` subproject. - Create `lightql-elastic-http` for ES REST materalizer implementation. - Add `show` method to ES materializers to transform a `lightQL` `Query` into the equivalent JSON. - Create `lightql-elastic-test` for testing both TCP & HTTP ES materializers
Update to elastic4s 5.5.3 and Fastparse 1.0.0
