+++ .travis.yml
@@ -4,7 +4,7 @@ sudo: required
 
 scala:
   - "2.11.11"
-  - "2.12.2"
+  - "2.12.3"
 
 jdk:
   - oraclejdk8
