+++ docs/src/main/tut/quickstart.md
@@ -5,17 +5,33 @@ to [Maven Central][maven-central] (albeit the latter is not an exact science). T
 cross-built both for Scala 2.11 and 2.12.
 
 
-In order to work with the Elasticsearch materializer (the only available for the time being),
+In order to work with either the HTTP or TCP Elasticsearch materializer (the only ones available for the time being),
 just include the proper SBT configuration[^1sbt]:
 
 ```scala
-libraryDependencies += "com.tecsisa" %% "lightql-elastic" % "5.4.2"
+libraryDependencies += "com.tecsisa" %% "lightql-elastic-http" % "5.5.0-SNAPSHOT"
+```
+or
+```scala
+libraryDependencies += "com.tecsisa" %% "lightql-elastic-tcp" % "5.5.0-SNAPSHOT"
 ```
 
 Then, open a SBT REPL session and try this sequence of commands:
 
+HTTP:
+
 ```tut
-import com.tecsisa.lightql.parser._, com.tecsisa.lightql.mat.elastic._
+import com.tecsisa.lightql.parser._, com.tecsisa.lightql.mat.elastic.http._
+
+val qs = "foo = 100"
+
+materialize(parse(qs).get.value)
+```
+
+TCP:
+
+```tut:reset
+import com.tecsisa.lightql.parser._, com.tecsisa.lightql.mat.elastic.tcp._
 
 val qs = "foo = 100"
 
@@ -25,8 +41,20 @@ materialize(parse(qs).get.value)
 On the other hand, if you're a happy [elastic4s][elastic4s-github-url] user (and sure you are) you'll be able to take advantage
 of the seamless integration between these two libraries and declare search expressions easily:
 
-```tut
-import com.tecsisa.lightql.parser._, com.tecsisa.lightql.mat.elastic._, com.sksamuel.elastic4s.ElasticDsl._
+HTTP:
+
+```tut:reset
+import com.tecsisa.lightql.parser._, com.tecsisa.lightql.mat.elastic.http._, com.sksamuel.elastic4s.http.ElasticDsl._
+
+def q(qs: String) = parse(qs).get.value
+
+search("songs") query q("composer = \"Johann Sebastian Bach\"")
+```
+
+TCP:
+
+```tut:reset
+import com.tecsisa.lightql.parser._, com.tecsisa.lightql.mat.elastic.tcp._, com.sksamuel.elastic4s.ElasticDsl._
 
 def q(qs: String) = parse(qs).get.value
